Clean up beta features storage code
